Packet decoding ,client handling and testing

+++ src/main/scala/rtmp/header/BasicHeaderDecoder.scala
@@ -8,7 +8,7 @@ import akka.util.ByteIterator
  */
 class BasicHeaderDecoder extends HeaderDecoder {
 
-  override def decode(firstByte: Byte, bufferItr: ByteIterator): Header = {
+  override def decode(firstByte: Int, bufferItr: ByteIterator): Header = {
 
     val sid = decodeSID(firstByte, bufferItr)
     BasicHeader(sid)
